+++ packages/components/dialog/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.1](https://github.com/mondaycom/vibe/compare/@vibe/dialog@4.0.0...@vibe/dialog@4.0.1) (2026-04-16)
+
+**Note:** Version bump only for package @vibe/dialog
+
+
+
+
+
 ## [3.0.12](https://github.com/mondaycom/vibe/compare/@vibe/dialog@3.0.11...@vibe/dialog@3.0.12) (2026-02-27)
 
 **Note:** Version bump only for package @vibe/dialog